+++ dist_node/lex/string_lexer.js
@@ -6,9 +6,9 @@
 var __o = require("bennu")["parse"],
     __o0 = require("bennu")["text"],
     __o1 = require("bennu")["lang"],
-    __o2 = require("nu-stream")["stream"],
-    __o3 = require("./line_terminator_lexer"),
-    __o4 = require("./number_lexer"),
+    __o2 = require("./line_terminator_lexer"),
+    __o3 = require("./number_lexer"),
+    __o4 = require("./common"),
     doubleQuote, escape, singleQuote, lineContinuation, unicodeEscapeSequence, hexEscapeSequence,
         characterEscapeSequence, escapeSequence, singleStringCharacter, singleStringCharacters, singleStringLiteral,
         doubleStringCharacter, doubleStringCharacters, doubleStringLiteral, stringLiteral, always = __o["always"],
@@ -27,17 +27,13 @@ var __o = require("bennu")["parse"],
     noneOf = __o0["noneOf"],
     between = __o1["between"],
     times = __o1["times"],
-    foldl = __o2["foldl"],
-    lineTerminatorSequence = __o3["lineTerminatorSequence"],
-    lineTerminator = __o3["lineTerminator"],
-    decimalDigit = __o4["decimalDigit"],
-    hexDigit = __o4["hexDigit"],
-    x, __add = (function(x, y) {
-        return (x + y);
-    }),
-    join = foldl.bind(null, __add, ""),
-    fromCharCode = map.bind(null, (function(x) {
-        return String.fromCharCode(parseInt(join(x), 16));
+    lineTerminatorSequence = __o2["lineTerminatorSequence"],
+    lineTerminator = __o2["lineTerminator"],
+    decimalDigit = __o3["decimalDigit"],
+    hexDigit = __o3["hexDigit"],
+    join = __o4["join"],
+    x, fromCharCode = map.bind(null, (function(x) {
+        return String.fromCharCode(parseInt(x, 16));
     }));
 (doubleQuote = character("\""));
 (singleQuote = character("'"));
@@ -59,21 +55,19 @@ var singleEscapeCharacter = oneOf("\"'\\bfnrtv")
     }))),
     escapeCharacter = choice(singleEscapeCharacter, decimalDigit, oneOf("ux")),
     nonEscapeCharacter = sequence(not(escapeCharacter), not(lineTerminator), anyToken);
-(hexEscapeSequence = next(character("x"), fromCharCode(times(2, hexDigit))));
-(unicodeEscapeSequence = next(character("u"), fromCharCode(times(4, hexDigit))));
+(hexEscapeSequence = next(character("x"), fromCharCode(join(times(2, hexDigit)))));
+(unicodeEscapeSequence = next(character("u"), fromCharCode(join(times(4, hexDigit)))));
 (characterEscapeSequence = either(singleEscapeCharacter, nonEscapeCharacter));
 (escapeSequence = choice(characterEscapeSequence, sequence(character("0"), not(decimalDigit), always("\u0000")),
     hexEscapeSequence, unicodeEscapeSequence));
 (singleStringCharacter = label("Single String Character", choice(attempt(lineContinuation), next(escape, escapeSequence),
     next(not(lineTerminator), noneOf("'\\")))));
 (singleStringCharacters = many(singleStringCharacter));
-(singleStringLiteral = label("Single String Literal", between(singleQuote, singleQuote, map(join,
-    singleStringCharacters))));
+(singleStringLiteral = label("Single String Literal", between(singleQuote, singleQuote, join(singleStringCharacters))));
 (doubleStringCharacter = choice(attempt(lineContinuation), next(escape, escapeSequence), next(not(lineTerminator),
     noneOf("\"\\"))));
 (doubleStringCharacters = many(doubleStringCharacter));
-(doubleStringLiteral = label("Double String Literal", between(doubleQuote, doubleQuote, map(join,
-    doubleStringCharacters))));
+(doubleStringLiteral = label("Double String Literal", between(doubleQuote, doubleQuote, join(doubleStringCharacters))));
 (stringLiteral = label("Sting Literal Lexer", either(singleStringLiteral, doubleStringLiteral)));
 (exports["doubleQuote"] = doubleQuote);
 (exports["escape"] = escape);