+++ dist_node/parse/package_parser.js
@@ -10,7 +10,8 @@ var __o = require("bennu")["parse"],
     __o2 = require("./token_parser"),
     __o3 = require("./value_parser"),
     __o4 = require("./statement_parser"),
-    khepriPackage, attempt = __o["attempt"],
+    __o5 = require("./pattern_parser"),
+    packageExport, packageExports, packageBody, khepriPackage, attempt = __o["attempt"],
     eager = __o["eager"],
     either = __o["either"],
     enumeration = __o["enumeration"],
@@ -26,19 +27,23 @@ var __o = require("bennu")["parse"],
     punctuator = __o2["punctuator"],
     identifier = __o3["identifier"],
     stringLiteral = __o3["stringLiteral"],
-    operator = __o3["operator"],
+    symbol = __o3["symbol"],
     withStatement = __o4["withStatement"],
     blockStatement = __o4["blockStatement"],
-    symbol = either(identifier, operator),
-    packageExport = label("Package Export", either(node(symbol, ast_package.PackageExport.create), nodea(enumeration(
-        stringLiteral, next(punctuator(":"), identifier)), (function(loc, alias, id) {
+    identifierPattern = __o5["identifierPattern"],
+    simplePackageExport = node(symbol, ast_package.PackageExport.create),
+    aliasedPackageExport = nodea(enumeration(either(stringLiteral, identifierPattern), next(punctuator(":", "#"),
+        identifier)), (function(loc, alias, id) {
         return ast_package.PackageExport.create(loc, id, alias);
-    })))),
-    packageExportList = label("Package Export List", node(between(punctuator("("), punctuator(")"), eager(sepBy(
-        optional(punctuator(",")), packageExport))), ast_package.PackageExports.create)),
-    packageExports = label("Package Exports", either(node(attempt(symbol), ast_package.PackageExport.create),
-        packageExportList)),
-    packageBody = label("Package Body", either(withStatement, blockStatement));
+    }));
+(packageExport = label("Package Export", either(attempt(aliasedPackageExport), simplePackageExport)));
+var packageExportList = label("Package Export List", node(between(punctuator("("), punctuator(")"), eager(sepBy(
+    optional(punctuator(",")), packageExport))), ast_package.PackageExports.create));
+(packageExports = label("Package Exports", either(attempt(simplePackageExport), packageExportList)));
+(packageBody = label("Package Body", either(withStatement, blockStatement)));
 (khepriPackage = label("Package", nodea(next(keyword("package"), enumeration(expected("package exports", packageExports),
     expected("package body", packageBody))), ast_package.Package.create)));
+(exports["packageExport"] = packageExport);
+(exports["packageExports"] = packageExports);
+(exports["packageBody"] = packageBody);
 (exports["khepriPackage"] = khepriPackage);