+++ dist_node/parse/parser.js
@@ -3,16 +3,14 @@
  * DO NOT EDIT
 */
 "use strict";
-var parse = require("bennu")["parse"],
-    __o = require("nu-stream")["stream"],
-    __o0 = require("khepri-ast")["position"],
+var __o = require("bennu")["parse"],
+    __o0 = require("nu-stream")["stream"],
     __o1 = require("./program_parser"),
-    parserStream, ParserPosition, parseInput, parseStream, ParserState = parse["ParserState"],
-    filter = __o["filter"],
-    isEmpty = __o["isEmpty"],
-    first = __o["first"],
-    streamFrom = __o["from"],
-    SourcePosition = __o0["SourcePosition"],
+    ParserPosition = require("./parser_position"),
+    parserStream, parseInput, parseStream, ParserState = __o["ParserState"],
+    runState = __o["runState"],
+    filter = __o0["filter"],
+    streamFrom = __o0["from"],
     program = __o1["program"];
 (parserStream = filter.bind(null, (function(x) {
     switch (x.type) {
@@ -24,36 +22,13 @@ var parse = require("bennu")["parse"],
             return true;
     }
 })));
-(ParserPosition = (function(tokenPosition, sourcePosition, prevEnd) {
-    var self = this;
-    (self.tokenPosition = tokenPosition);
-    (self.sourcePosition = sourcePosition);
-    (self.prevEnd = prevEnd);
-}));
-(ParserPosition.prototype = new(parse.Position)());
-(ParserPosition.prototype.constructor = ParserPosition);
-(ParserPosition.initial = new(ParserPosition)(parse.Position.initial, SourcePosition.initial, parse.Position.initial));
-(ParserPosition.prototype.increment = (function(tok, r) {
-    var self = this;
-    return new(ParserPosition)(self.tokenPosition.increment(tok), (isEmpty(r) ? tok.loc.end : first(r)
-        .loc.start), tok.loc.end);
-}));
-(ParserPosition.prototype.toString = (function() {
-    var self = this;
-    return ("" + self.sourcePosition);
-}));
-(ParserPosition.prototype.compare = (function(pos) {
-    var self = this;
-    return self.tokenPosition.compare(pos.tokenPosition);
-}));
 (parseStream = (function(s) {
-    return parse.runState(program, new(ParserState)(parserStream(s), ParserPosition.initial));
+    return runState(program, new(ParserState)(parserStream(s), ParserPosition.initial));
 }));
 (parseInput = (function(z) {
     var s = streamFrom(z);
-    return parse.runState(program, new(ParserState)(parserStream(s), ParserPosition.initial));
+    return runState(program, new(ParserState)(parserStream(s), ParserPosition.initial));
 }));
 (exports["parserStream"] = parserStream);
-(exports["ParserPosition"] = ParserPosition);
 (exports["parseInput"] = parseInput);
 (exports["parseStream"] = parseStream);